refactor: Servlet 클래스들을 /web/servlet 하위로 이동
