+++ src/main/java/woowa/camp/jspcafe/web/servlet/ArticleWriteServlet.java
@@ -1,4 +1,4 @@
-package woowa.camp.jspcafe.web;
+package woowa.camp.jspcafe.web.servlet;
 
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletConfig;
